use helper get_azure_ad_token_from_entrata_id
